requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
opencv-python
Pillow
imgaug
Flask
torch
torchvision
fastapi
uvicorn
opencv-python
pillow